Move basic functionalities from core to base module
